refactor actionFooter actions to utils file, added currentDoc to utils file
