+++ src/components/utils/currentDoc.js
@@ -1,0 +1,1 @@
+export const doc = 'DGtOTuQcZfamL6ZvJjTn';
