+++ src/components/utils/index.js
@@ -1,0 +1,2 @@
+export * from './actionFooter';
+export { doc } from './currentDoc';
